+++ app/page.tsx
@@ -1,4 +1,3 @@
-"use client";
 import About from "@/components/about";
 import Contact from "@/components/contact";
 import Hero from "@/components/hero";
