+++ components/nav.tsx
@@ -1,9 +1,11 @@
 "use client";
 import Link from "next/link";
 import { useState } from "react";
+import { navList } from "@/lib/data";
+
 import ThemeSwitcher from "./theme-switcher";
 import Logo from "./logo";
-import { navList } from "@/lib/data";
+import Menu from "./nav-menu";
 
 export default function Nav() {
   let [mobileMenu, setMobileMenu] = useState(false);
@@ -17,50 +19,8 @@ export default function Nav() {
       <div className="max-w-screen-xl flex flex-wrap items-center justify-between mx-auto p-4">
         <Logo />
         <div className="flex md:order-2 space-x-3 md:space-x-0 rtl:space-x-reverse">
-          {<ThemeSwitcher />}
-          <button
-            data-collapse-toggle="navbar-cta"
-            type="button"
-            className="inline-flex items-center p-2 w-10 h-10 justify-center text-sm text-gray-500 rounded-lg md:hidden hover:bg-gray-100 focus:outline-none dark:text-gray-400 dark:hover:bg-gray-700"
-            aria-controls="navbar-cta"
-            aria-expanded="false"
-            onClick={toggleMobileMenu}
-          >
-            <span className="sr-only">Open main menu</span>
-            {mobileMenu ? (
-              <svg
-                className="w-5 h-5"
-                aria-hidden="true"
-                xmlns="http://www.w3.org/2000/svg"
-                fill="none"
-                viewBox="0 0 24 24"
-              >
-                <path
-                  stroke="currentColor"
-                  strokeLinecap="round"
-                  strokeLinejoin="round"
-                  strokeWidth="2"
-                  d="M6 18L18 6M6 6l12 12"
-                />
-              </svg>
-            ) : (
-              <svg
-                className="w-5 h-5"
-                aria-hidden="true"
-                xmlns="http://www.w3.org/2000/svg"
-                fill="none"
-                viewBox="0 0 17 14"
-              >
-                <path
-                  stroke="currentColor"
-                  strokeLinecap="round"
-                  strokeLinejoin="round"
-                  strokeWidth="2"
-                  d="M1 1h15M1 7h15M1 13h15"
-                />
-              </svg>
-            )}
-          </button>
+          <ThemeSwitcher />
+          <Menu mobileMenu={mobileMenu} toggleMobileMenu={toggleMobileMenu} />
         </div>
         <div
           className={`${
@@ -74,7 +34,7 @@ export default function Nav() {
                 <li key={item.name}>
                   <Link
                     href={item.link}
-                    className="block text-lg py-2 px-3 text-gray-900 rounded hover:bg-gray-100 md:hover:bg-transparent md:border-0 md:hover:text-blue-700 md:p-0 dark:text-white md:dark:hover:text-blue-500 dark:hover:bg-gray-700 dark:hover:text-white md:dark:hover:bg-transparent"
+                    className="block text-sm font-semibold uppercase py-2 px-3 text-gray-900 rounded hover:bg-gray-200 md:hover:bg-transparent md:border-0 md:hover:text-violet-700 md:p-0 dark:text-white md:dark:hover:text-violet-500 dark:hover:bg-gray-700 dark:hover:text-white md:dark:hover:bg-transparent"
                   >
                     {item.name}
                   </Link>
