+++ lib/data.js
@@ -75,3 +75,9 @@ export const navList = [
   { name: "Projects", link: "/#projects" },
   { name: "Contact", link: "/#contact" },
 ];
+
+export const contact = {
+  heading: "Contact",
+  subHeading:
+    "Feel free to Contact me by submitting the form below and I will get back to you as soon as possible",
+};
